feat(transpiler): map TypeScript `number` to Udon Double for JS-consistent semantics
JavaScript's `number` is IEEE 754 double-precision, but the transpiler was
mapping it to Udon Single (32-bit float). This caused precision loss for
literals and pushed callers into ambiguous territory when stub APIs declared
`number` for what should be either Single or Double.

Switch the default to Double end-to-end and patch the latent integer-width
assumptions that surfaced as a result.

Type mapping (Phase A core):
- `type_mapper.ts`: BUILTIN_NAME_MAP + inferLiteralType -> Double
- `type_checker_type_resolver.ts`: NumberLiteral / NumberLike -> Double
- `parser/types.ts`: NumericLiteral / PrefixUnary fallback -> Double
- `parser/visitors/declaration.ts`: untyped param/property fallback -> Double
- `parser/errors.ts`: error LiteralNode placeholder -> Double
- `udon_type_resolver.ts`: TS_PRIMITIVE_ALIASES `number` -> System.Double
- `optimizer/passes/constant_folding.ts`: numeric folding fallback -> Double
- `ir/converter.ts`: newTemp default -> Double (no-arg callers: 0)
- `ir/inline.ts`: DataList sentinel 0 -> Double (read-only sentinel)
- `ir/expression.ts:visitArrayAccessExpression`: DataList valueType
  fallback -> Double

Stub disambiguation:
- `DataContainerTypes.ts`: DataToken accessors switch from `number` to
  explicit brand types — `Float!: UdonFloat`, `Double!: UdonDouble`,
  `Long!: UdonLong`, `TokenType!: UdonInt`. Without this the metadata
  registry resolves both Float and Double to System.Double after the
  remap, emitting `__get_Float__SystemDouble` (bogus) at every read.

Width-mismatch fixes (Phase A surfaced these):
- `ir/visitors/statement.ts:visitVariableDeclaration`: extend the
  Object/Single auto-narrow to also cover Double. Without this,
  `let x = "abc".indexOf(...)` declares a Double slot but the COPY
  source is Int32 — the upper 4 bytes of the Double slot are uninitialised
  garbage and downstream Double comparisons crash.
- `ir/visitors/call.ts:visitMathStaticCall`: dispatch Math.floor/ceil/
  abs/round/etc. to `SystemMath.__<Op>__SystemDouble__SystemDouble` when
  any argument is Double. Mathf operates on Single; pushing 8 bytes into
  the 4-byte input slot reads garbage.
- `ir/visitors/expression.ts:visitShortCircuitOr`: also reject
  `udonType === Object` from the value-typed expectedType filter.
  Inline expansion can leak a `ClassTypeSymbol` whose `udonType` is Object
  (e.g. an inlined return target), and the previous reference-equality
  check `!== ObjectType` did not catch it. The bool result of `a || b`
  was then COPY'd into an %SystemObject slot, corrupting downstream reads
  (visible as 34 successive "Invalid TileKind" cascades from
  `Tile.fromKind`).

Safer literal demotion + DataToken fail-safe:
- `ir/visitors/expression.ts:visitLiteral`: demote integer-valued
  number literals to the integer expected type only when the value
  fits the target's range (Byte/SByte/Int16/UInt16/Int32/UInt32 ranges
  enumerated). bigint literals stay Int64. Out-of-range values fall
  through to the regular Cast pipeline (e.g. `256 as UdonByte` keeps
  the SystemConvert path).
- `ir/helpers/assignment.ts:wrapDataToken`: route both Single AND
  Double through `DataToken.op_Implicit` instead of the ctor. The
  Single ctor is registered but throws at runtime on the current SDK;
  Double ctor is untested on real hardware so the implicit operator
  stays the safer path.

Tests:
- Mechanical update of `SystemSingle*` / `SystemSingleArray` expectations
  to `SystemDouble*` / `SystemDoubleArray` in fixtures that exercise
  default `number` (~14 test files). UdonFloat-explicit tests keep
  Single. `Bug 15` (any-escape Single→Int32 setter) is resolved by the
  remap and op_Implicit fail-safe; promoted from `it.fails` to `it`.

Verification:
- `pnpm typecheck`: clean
- `pnpm test --run`: 861 passed / 0 failed (1044 with skips)
- VM minimal repros (DataToken Double ctor / op_Implicit / IndexOf /
  Math.floor / Single→Double conversion): all pass on real Unity
- VM mahjong-t2 integration: 12/38 pass (parity with baseline,
  no regression). Remaining 26 failures are downstream of
  `D3 method dispatch miss: check on untracked instance` — a known
  inline-handle dispatch limitation independent of this remap.

+++ src/transpiler/frontend/parser/errors.ts
@@ -77,7 +77,7 @@ export function createUnsupportedExpressionPlaceholder(
   const result: LiteralNode = {
     kind: ASTNodeKind.Literal,
     value: 0,
-    type: PrimitiveTypes.single,
+    type: PrimitiveTypes.double,
   };
   return tsNode ? this.attachLoc(tsNode, result) : result;
 }
